Removes workaround for eslint in katello
We added this workaround due to '.eslintrc' being
"up the tree" for angular tests run in  engines/
bastion_katello. The root .eslintrc is for
React code. A Katello PR will move the .eslintrc
file to webpack/ so its no longer "up in the tree"
and picked up by running eslint in engines/ directory