Use immutable variables
In a for(x in y) the variable x is mutable. That means its value changes
during the loop. The repoclosure() call uses this variable and thus
results in the last value. This iteration doesn't have that issue.

It also fixes an issue where the push loop used the incorrect variable.