Rewrite Smart Proxy PR testing to a pipeline
The main motivation for this was to get rid of the checkstyle plugin. It
now rewrites to warnings-ng.

To keep things easy, it gets rid of the jenkins:rubocop task. This means
the task will work on all branches without updating them all.

It only runs Rubocop on Ruby 2.7 because recordIssues has problems when
it runs multiple times. This has the side effect of not running the
tests on 2.7 when Rubocop fails. This means the number of tests
fluctuates.