build plugins on bullseye not focal
Bundler on focal is broken and doesn't find thor properly.
While we have fixes for that for the built packages, it is still a
problem during building of plugins (as this requires to load the Bundler
library vs using the CLI).

The setup on Bullseye is sane and seems to just work from my tests, so
let's use that. Bullseye should be able to build packages for Foreman
3.2+, which is everything we support these days.