don't put individual playbooks in own stages
Jenkins doesn't cope well with parallel stages inside stages when we
dynamically generate the list of steps.

If we'd use just `parallel` or `matrix` with a static set of steps, it
would work, but we generate things on the fly and this confuses the Blue
Ocean view, resulting in only the first stage to be present in the view
and thus making it impossible to jump to the failed step if there are
any.