Refactor foreman-packaging-*-release jobs for reuse
This introduces a completely reusable
foreman-packaging-{packaging_label}-{version}-release job template. Then
it instantiates the various versions of it.

The design is currently that it creates both koji and copr for nightly,
but only a single rpm instance for releases.

The branch-foreman script is modified to do the right thing.