inject foreman version for stable plugin builds
this allows to have the same plugin version to have different builds for
different foreman releases, thus making it possible to cleanly upgrade
to plugin builds that match the foreman release even if there are no
plugin version differences.

practical example: foreman_puppet 6.2.0 exists in 3.8, 3.9, 3.10 and
nightly, but in 3.10 and nightly it needs to be built with Webpack5 and
not 3 to work.
before this change, all builds ended up being 6.2.0-1 and apt would not
consider to upgrade those when coming from 3.9 to 3.10, creating a
broken setup.
with this change, the versions would be:
6.2.0-1~fm3.8
6.2.0-1~fm3.9
6.2.0-1~fm3.10
6.2.0-1
which apt will gladly upgrade in that order

It would be cool if we could name the last one 6.2.0-1~3.11, but there
is no way for us to know that the next version will be 3.11 at that
point